refactor: handle nextauth signin error, extract useGoogleOAuthCallback hook, add strict mode tests
